Update README.md
Please note that I uploaded the data twice initially. All of the answer screenshots that aren't the same as the challenge answers are exactly double because the search returns twice the events. Also, for one of the questions dealing with purchases I used "status=200" even though the results of the class screenshots did not (compared to my answers, using no "status=200" returned the expected answer in the screenshot) as a different status would not be considered a successful purchase.